Example tableone Object with Imaginary Demographic Data for Practicing with the tablestylizer Package

This dataset is an example tableone object created using the tableone package, representing a summary table of characteristics. It is generated based on the dataset "tablestylizer_df" and is in the format of a list with components for continuous (ContTable) and categorical (CatTable) variables. Users can practice and explore the functionality of the tablestylizer package by tidying up this table with the tablestylizer::stylize_tableone() function. Note that in this table, all data is presented under the "Overall" column, without dividing into Groups A, B, and C.

Usage

tableone_1g

Format

An example tableone object with the following components:

ContTable

Summary statistics for continuous variables (Age (years), Height (inches)).

CatTable

Summary statistics for categorical variables (Women, Likes coffee, Favorite color, Group).

MetaData

Metadata information, including variable types and labels.

Source

This data is generated for the purpose of providing a practice example tableone object for learning and experimenting with the tablestylizer package. It does not represent real-world individuals or situations.

Examples

tableone_1g
#>                              
#>                               Overall      
#>   n                              20        
#>   Age (years) (mean (SD))     51.05 (21.24)
#>   Women = Yes (%)                11 (55.0) 
#>   Height (inches) (mean (SD)) 68.30 (2.23) 
#>   Likes coffee = Yes (%)          8 (40.0) 
#>   Favorite color (%)                       
#>      Red                          6 (30.0) 
#>      Green                        7 (35.0) 
#>      Blue                         7 (35.0) 
#>   Group (%)                                
#>      A                            5 (25.0) 
#>      B                            8 (40.0) 
#>      C                            7 (35.0)