Example Dataset with Imaginary Demographic Data for Practicing with the tablestylizer Package

This dataset provides an example of fictional data related to a group of individuals. It includes information such as age, sex, height, coffee preferences, favorite color, and group affiliation.

Usage

tablestylizer_df

Format

A tibble with 20 rows and 7 variables:

ID

Unique identifier for each individual.

Age (years)

Age of the individuals in years.

Women

Indicates whether the individual is female (Yes) or not (No).

Height (inches)

Height of the individuals in inches.

Likes coffee

Individuals' preference for coffee (Yes or No).

Favorite color

Favorite color of the individuals (Red, Green, or Blue).

Group

Group affiliation of the individuals (A, B, or C).

Source

This data is generated for the purpose of illustrating data manipulation and analysis in R. It does not represent real-world individuals or situations.

Examples

head(tablestylizer_df)
#> # A tibble: 6 × 7
#>   ID    `Age (years)` Women `Height (inches)` `Likes coffee` `Favorite color`
#>   <chr>         <int> <fct>             <dbl> <fct>          <fct>           
#> 1 01               41 Yes                  67 No             Blue            
#> 2 02               18 Yes                  67 No             Blue            
#> 3 03               21 Yes                  67 No             Green           
#> 4 04               73 Yes                  65 No             Red             
#> 5 05               43 No                   69 No             Red             
#> 6 06               70 Yes                  66 Yes            Green           
#> # ℹ 1 more variable: Group <fct>